clean up files made safer

+++ src/main/java/edu/harvard/iq/dataverse/dataaccess/FileAccessIO.java
@@ -726,8 +726,11 @@ public class FileAccessIO<T extends DvObject> extends StorageIO<T> {
     }
 
     @Override
-    public List<String> cleanUp(Predicate<String> filter) throws IOException {
+    public List<String> cleanUp(Predicate<String> filter, boolean dryRun) throws IOException {
         List<String> toDelete = this.listAllFiles().stream().filter(filter).collect(Collectors.toList());
+        if (dryRun) {
+            return toDelete;
+        }
         for (String f : toDelete) {
             this.deleteFile(f);
         }
