+++ src/main/java/edu/harvard/iq/dataverse/dataaccess/S3AccessIO.java
@@ -1309,8 +1309,6 @@ public class S3AccessIO<T extends DvObject> extends StorageIO<T> {
         return true;
     }
     
-
-
     private List<String> listAllFiles() throws IOException {
         if (!this.canWrite()) {
             open();
@@ -1372,8 +1370,11 @@ public class S3AccessIO<T extends DvObject> extends StorageIO<T> {
     }
 
     @Override
-    public List<String> cleanUp(Predicate<String> filter) throws IOException {
+    public List<String> cleanUp(Predicate<String> filter, boolean dryRun) throws IOException {
         List<String> toDelete = this.listAllFiles().stream().filter(filter).collect(Collectors.toList());
+        if (dryRun) {
+            return toDelete;
+        }
         for (String f : toDelete) {
             this.deleteFile(f);
         }
